+++ src/lib/pages/account-details/components/delegations/RedelegationsSection.tsx
@@ -39,7 +39,7 @@ export const RedelegationsSection = ({
       />
     </Flex>
     {isLoading ? (
-      <Loading />
+      <Loading withBorder />
     ) : (
       <>
         <NonRedelegatableSection redelegations={redelegations} />