+++ src/lib/pages/block-details/index.tsx
@@ -20,7 +20,7 @@ const BlockDetail = () => {
     if (router.isReady) AmpTrack(AmpEvent.TO_BLOCK_DETAIL);
   }, [router.isReady]);
 
-  if (isLoading) return <Loading />;
+  if (isLoading) return <Loading withBorder />;
 
   return (
     <PageContainer>